requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
pandas==2.2.0
openpyxl==3.1.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose==3.3.0
aiofiles==23.1.0
jinja2==3.1.2
filelock==3.12.4
aiofiles==23.1.0
python-dotenv==1.0.1